Refactoring MU controller for reuse in organisation page #2938
