+++ grails-app/conf/spring/resources.groovy
@@ -2,6 +2,7 @@
 import au.org.ala.merit.hub.HubAwareLinkGenerator
 import au.org.ala.merit.StatisticsFactory
 import au.org.ala.merit.CheckRisksAndThreatsTask
+import au.org.ala.merit.util.ProjectGroupingHelper
 
 // Place your Spring DSL code here
 beans = {
@@ -14,4 +15,6 @@ beans = {
     statisticsFactory(StatisticsFactory)
 
     checkForRisksAndThreatsTask(CheckRisksAndThreatsTask)
+
+    projectGroupingHelper(ProjectGroupingHelper)
 }
